Clean code && fix bugs (#2)
* pre format money

* clean code

* portrait orientation only

* no more need

* use big decimal

* clean code

* unit test for formatter

* unit test for formatter

* formatter was rewrited

* load rates from net of base rate > 0

* fix one more case

* zero without floating point for rates

* fix rx sequence when filter for zero base rate value

* fix one more case for formatter

* release 1.0.1

+++ app/src/main/java/ru/bagrusss/revolutdemo/screens/rates/models/Rate.kt
@@ -1,5 +1,7 @@
 package ru.bagrusss.revolutdemo.screens.rates.models
 
+import java.math.BigDecimal
+
 /**
  * Created by bagrusss on 13.08.2019
  */
@@ -8,5 +10,5 @@ class Rate(
     @JvmField val title: String,
     @JvmField val description: String,
     @JvmField val imgUrl: String,
-    @JvmField val cost: Double
+    @JvmField val cost: BigDecimal
 )
